+++ COMMANDS/ability.lua
@@ -2,7 +2,7 @@
 local S = minetest.get_translator(minetest.get_current_modname())
 
 better_commands.register_command("ability", {
-    params = "<player> <priv> [value]",
+    params = S("<player> <priv> [value]"),
     description = S("Sets <priv> of <player> to [value] (true/false). If [value] is not supplied, returns the existing value of <priv>"),
     privs = {privs = true},
     func = function(name, param, context)