+++ COMMANDS/chat.lua
@@ -2,7 +2,7 @@
 local S = minetest.get_translator(minetest.get_current_modname())
 
 better_commands.register_command("say", {
-    params = "<message>",
+    params = S("<message>"),
     description = S("Says <message> to all players (which can include selectors such as @@a if you have the server priv)"),
     privs = {shout = true},
     func = function(name, param, context)
@@ -25,7 +25,7 @@ better_commands.register_command("say", {
 })
 
 better_commands.register_command("msg", {
-    params = "<target> <message>",
+    params = S("<target> <message>"),
     description = S("Sends <message> privately to <target> (which can include selectors like @@a if you have the server priv)"),
     privs = {shout = true},
     func = function(name, param, context)
@@ -66,7 +66,7 @@ better_commands.register_command_alias("tell", "msg")
 
 better_commands.register_command("me", {
     description = S("Broadcasts a message about yourself (which can include selectors like @@a if you have the server priv)"),
-    params = "<action>",
+    params = S("<action>"),
     privs = {shout = true},
     func = function(name, param, context)
         context = better_commands.complete_context(name, context)
@@ -86,3 +86,51 @@ better_commands.register_command("me", {
         return true, nil, 1
     end
 })
+
+better_commands.register_command("teammsg", {
+    params = S("<message>"),
+    description = S("Sends <message> privately to all team members (which can include selectors like @@a if you have the server priv)"),
+    privs = {shout = true},
+    func = function(name, param, context)
+        context = better_commands.complete_context(name, context)
+        if not context then return false, S("Missing context"), 0 end
+        if not context.executor then return false, S("Missing executor"), 0 end
+        local split_param = better_commands.parse_params(param)
+        if not split_param[1] then
+            return false, nil, 0
+        end
+        if not (context.executor.is_player and context.executor:is_player()) then
+            return false, S("/teammsg can only be used by players"), 0
+        end
+        local sender = context.executor:get_player_name()
+        local team = better_commands.teams.players[sender]
+        local team_color = better_commands.team_colors[better_commands.teams.teams[team].color or "white"]
+        local display_name = better_commands.teams.teams[team].display_name or team
+        if not team then return false, S("/teammsg cannot be used unless on a team"), 0 end
+        local start = S("[@1] <@2> ", minetest.colorize(team_color, display_name), better_commands.get_entity_name(context.executor))
+        local message
+        if context.command_block or minetest.check_player_privs(context.origin, {server = true}) then
+            local err
+            message, err = better_commands.expand_selectors(param, split_param, 1, context)
+            if err then return false, err, 0 end
+        else
+---@diagnostic disable-next-line: param-type-mismatch
+            message = param:sub(split_param[1][1], -1)
+        end
+        local count = 0
+        minetest.chat_send_player(name, "-> "..start..message)
+        for receiver, receiverteam in pairs(better_commands.teams.players) do
+            if receiverteam == team then
+                count = count + 1
+                if minetest.get_player_by_name(receiver) then
+                    if receiver ~= name then
+                        minetest.chat_send_player(receiver, start..message)
+                    end
+                end
+            end
+        end
+        return true, nil, count
+    end
+})
+
+better_commands.register_command_alias("tm", "teammsg")