+++ COMMANDS/playsound.lua
@@ -2,8 +2,8 @@
 local S = minetest.get_translator(minetest.get_current_modname())
 
 better_commands.register_command("playsound", {
-    params = "<sound> <targets|pos> [volume] [pitch] [maxDistance]",
-    description = "Plays a sound",
+    params = S("<sound> <targets|pos> [volume] [pitch] [maxDistance]"),
+    description = S("Plays a sound"),
     privs = {server = true},
     func = function(name, param, context)
         context = better_commands.complete_context(name, context)