+++ CHANGELOG.md
@@ -18,12 +18,14 @@ Initial release. Missing *lots* of commands, several `execute` subcommands, lots
   * `dug.<itemstring>` (same as `mined`)
   * `placed.<itemstring>`
   * `crafted.<itemstring>`
+  * `awards.<award>` (requires Awards mod)
 * Added `ascending|descending` argument for `/scoreboard objectives setdisplay`
 * Added `/gamemode` command (grants/revokes `creative` priv in non-MCL games)
 * Added `/spawnpoint` and `/clearspawnpoint` commands
 * Added `gamemode`/`m` selector argument
 * Added `level`/`l`/`lm` selector arguments
 * Added `/clear` command
+* Added `/teammsg` command
 ### Bugfixes
 * The `/kill` command is more likely to successfully kill entities.
 * The `rm`/`r` selector arguments now actually treat their values as numbers (not strings), and are now inclusive as intended.